fix comments: extract verification request mocks and add error handling to stats
