+++ server/verification-request/verification-request-stats.service.spec.ts
@@ -4,30 +4,7 @@ import { Test, TestingModule } from "@nestjs/testing";
 import { VerificationRequestService } from "./verification-request.service";
 import { VerificationRequestStatsService } from "./verification-request-stats.service";
 import { VerificationRequestDocument } from "./schemas/verification-request.schema";
-
-const createFakeVerificationRequest = (
-  overrides?: Partial<VerificationRequestDocument>
-) => ({
-  _id: "60c72b2f9f1b2c3d4e5f6a7b",
-  status: "In Triage",
-  sourceChannel: "Whatsapp",
-  data_hash: "AABBCCDD1122334455667788",
-  updatedAt: new Date(),
-  ...overrides,
-});
-
-const mockQuery = {
-  sort: jest.fn().mockReturnThis(),
-  limit: jest.fn().mockReturnThis(),
-  select: jest.fn().mockReturnThis(),
-  lean: jest.fn().mockReturnThis(),
-  exec: jest.fn(),
-};
-
-export const mockVerificationRequestModel = {
-  find: jest.fn().mockReturnValue(mockQuery),
-  aggregate: jest.fn().mockReturnThis(),
-};
+import { createFakeVerificationRequest, mockQuery, mockVerificationRequestModel } from "../mocks/VerificationRequestMock";
 
 describe("VerificationRequestStatsService (Unit)", () => {
   let testingModule: TestingModule;
